+++ test/_init.js
@@ -1,0 +1,8 @@
+// remove all orginal methods that should be polyfilled if missing
+String.prototype.includes = undefined;
+Object.assign = undefined;
+Number.isInteger = undefined;
+Math.log10 = undefined;
+
+// require 
+require('../dist/finka');